refactor: move existing query utils to a separate file
